+++ src/super_gradients/common/crash_handler/__init__.py
@@ -1,9 +1,4 @@
 from super_gradients.common.crash_handler.crash_handler import setup_crash_handler
-from super_gradients.common.crash_handler.exception import ExceptionInfo
-from super_gradients.common.crash_handler.crash_tips import get_relevant_crash_tip_message
 
 
-__all__ = ["setup_crash_handler", "get_relevant_crash_tip_message", "ExceptionInfo"]
-
-
-setup_crash_handler()
+__all__ = ["setup_crash_handler"]
